requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydub==0.25.1
elevenlabs==0.2.24
openai==1.3.0
python-multipart==0.0.6
pydantic==1.10.13
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.1 